examples/name_change/pyproject.toml
[project]
name = "name-change-example"
version = "0.1.0"
description = "Name change agent example using deepagents package"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=1.0.2",
    "langchain-anthropic>=1.0.3",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
    "deepagents>=0.2.6",
    "python-dotenv>=1.0.0",
    "langgraph-cli[inmem]>=0.1.55",
    "langgraph>=1.0.10",
]

[tool.uv]
override-dependencies = [
    "nbconvert>=7.17.0",  # CVE-2025-53000
    "protobuf>=6.33.5",  # CVE-2026-0994
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["name_change_agent"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
]